+++ pyeee/general_functions.py
@@ -8,15 +8,21 @@ The current functions are:
     logistic              Logistic function L/(1+exp(-k(x-x0)))
     logistic_p            logistic(x,*p)
     dlogistic             First derivative of logistic function
+    dlogistic_p           dlogistic(x,*p)
     d2logistic            Second derivative of logistic function
+    d2logistic_p          d2logistic(x,*p)
     logistic_offset       logistic function with offset L/(1+exp(-k(x-x0))) + a
     logistic_offset_p     logistic_offset(x,*p)
     dlogistic_offset      First derivative of logistic function with offset
+    dlogistic_offset_p    dlogistic_offset(x,*p)
     d2logistic_offset     Second derivative of logistic function with offset
+    d2logistic_offset_p   d2logistic_offset(x,*p)
     logistic2_offset      Double logistic function with offset L1/(1+exp(-k1(x-x01))) - L2/(1+exp(-k2(x-x02))) + a2
     logistic2_offset_p    logistic2_offset(x,*p)
     dlogistic2_offset     First derivative of double logistic function with offset
+    dlogistic2_offset_p   dlogistic2_offset(x,*p)
     d2logistic2_offset    Second derivative of double logistic function with offset
+    d2logistic2_offset_p  d2logistic2_offset(x,*p)
 
 This module was written by Matthias Cuntz while at Department of
 Computational Hydrosystems, Helmholtz Centre for Environmental
@@ -31,33 +37,46 @@ Released under the MIT License; see LICENSE file for details.
 * Added functions logistic_p and logistic_offset_p, Dec 2017, Matthias Cuntz
 * Changed to Sphinx docstring and numpydoc, Dec 2019, Matthias Cuntz
 * Distinguish iterable and array_like parameter types, Jan 2020, Matthias Cuntz
+* Make systematically function_p versions of all logistic functions and its derivatives, Feb 2020, Matthias Cuntz
 
 .. moduleauthor:: Matthias Cuntz
 
-The following wrappers are provided
+The following functions are provided
 
 .. autosummary::
     curvature
     logistic
     logistic_p
     dlogistic
+    dlogistic_p
     d2logistic
+    d2logistic_p
     logistic_offset
     logistic_offset_p
     dlogistic_offset
+    dlogistic_offset_p
     d2logistic_offset
+    d2logistic_offset_p
     logistic2_offset
     logistic2_offset_p
     dlogistic2_offset
+    dlogistic2_offset_p
     d2logistic2_offset
+    d2logistic2_offset_p
 """
 import numpy as np
 import scipy.special as sp
 
 __all__ = ['curvature',
-           'logistic', 'logistic_p', 'dlogistic', 'd2logistic',
-           'logistic_offset', 'logistic_offset_p', 'dlogistic_offset', 'd2logistic_offset',
-           'logistic2_offset', 'logistic2_offset_p', 'dlogistic2_offset', 'd2logistic2_offset']
+           'logistic', 'logistic_p',
+           'dlogistic', 'dlogistic_p',
+           'd2logistic', 'd2logistic_p',
+           'logistic_offset', 'logistic_offset_p',
+           'dlogistic_offset', 'dlogistic_offset_p',
+           'd2logistic_offset', 'd2logistic_offset_p',
+           'logistic2_offset', 'logistic2_offset_p',
+           'dlogistic2_offset', 'dlogistic2_offset_p',
+           'd2logistic2_offset', 'd2logistic2_offset_p']
 
 # -----------------------------------------------------------
 # curvature of function
@@ -85,7 +104,7 @@ def curvature(x, dfunc, d2func, *args, **kwargs):
     float or ndarray
         Curvature of function f at `x`
     """
-    return ( d2func(x, *args, ** kwargs) /
+    return ( d2func(x, *args, **kwargs) /
                  (1. + dfunc(x, *args, **kwargs)**2)**1.5 )
 
 
@@ -118,27 +137,9 @@ def logistic(x, L, k, x0):
 
 def logistic_p(x, p):
     """
-    Logistic function:
-
-        `L/(1+exp(-k(x-x0)))`
-
-    where parameters `L`, `k` and `x0` are given as one iterable, i.e.
-
-        `p[0]/(1+exp(-p[1](x-p[2])))`
-
-    Parameters
-    ----------
-    x : array_like
-        Independent variable to evalute logistic function
-    p : iterable
-        Iterable of length 3 with maximum, steepness and inflection point of logistic function
-
-    Returns
-    -------
-    float or ndarray
-        Logistic function at `x` with maximum `p[0]`, steepness `p[1]` and inflection point `p[2]`
+    Wrapper function for :func:`~pyeee.general_functions.logistic`: `logistic(x, *p)`.
     """
-    return logistic(x, p[0], p[1], p[2])
+    return logistic(x, *p)
 
 
 # -----------------------------------------------------------
@@ -170,6 +171,13 @@ def dlogistic(x, L, k, x0):
         First derivative of logistic function at `x` with maximum `L`, steepness `k` and inflection point `x0`
     """
     return k * L / (2. * (np.cosh(k * (x - x0)) + 1.))
+
+
+def dlogistic_p(x, p):
+    """
+    Wrapper function for :func:`~pyeee.general_functions.dlogistic`: `dlogistic(x, *p)`.
+    """
+    return dlogistic(x, *p)
 
 
 # -----------------------------------------------------------
@@ -204,6 +212,13 @@ def d2logistic(x, L, k, x0):
                  (2. * (np.cosh(k * (x - x0)) + 1.)**2) )
 
 
+def d2logistic_p(x, p):
+    """
+    Wrapper function for :func:`~pyeee.general_functions.d2logistic`: `d2logistic(x, *p)`.
+    """
+    return d2logistic(x, *p)
+
+
 # -----------------------------------------------------------
 # L/(1+exp(-k(x-x0))) + a - logistic function with offset
 def logistic_offset(x, L, k, x0, a):
@@ -235,27 +250,9 @@ def logistic_offset(x, L, k, x0, a):
 
 def logistic_offset_p(x, p):
     """
-    Logistic function with offset:
-
-        `L/(1+exp(-k(x-x0))) + a`
-
-    where parameters `L`, `k`, `x0`, and `a` are given as one iterable, i.e.
-
-        `p[0]/(1+exp(-p[1](x-p[2]))) + p[3]`
-
-    Parameters
-    ----------
-    x : array_like
-        Independent variable to evalute logistic function
-    p : iterable
-        Iterable of length 4 with maximum, steepness, inflection point, and offset of logistic function
-
-    Returns
-    -------
-    float or ndarray
-        Logistic function at `x` with maximum `p[0]`, steepness `p[1]`, inflection point `p[2]`, and offset `p[3]`
+    Wrapper function for :func:`~pyeee.general_functions.logistic_offset`: `logistic_offset(x, *p)`.
     """
-    return logistic_offset(x, p[0], p[1], p[2], p[3])
+    return logistic_offset(x, *p)
 
 
 # -----------------------------------------------------------
@@ -292,6 +289,13 @@ def dlogistic_offset(x, L, k, x0, a):
     return k * L / (2. * (np.cosh(k * (x - x0)) + 1.))
 
 
+def dlogistic_offset_p(x, p):
+    """
+    Wrapper function for :func:`~pyeee.general_functions.dlogistic_offset`: `dlogistic_offset(x, *p)`.
+    """
+    return dlogistic_offset(x, *p)
+
+
 # -----------------------------------------------------------
 # 2nd derivative of logistic functions with offset
 def d2logistic_offset(x, L, k, x0, a):
@@ -325,6 +329,13 @@ def d2logistic_offset(x, L, k, x0, a):
     """
     return ( -k**2 * L * np.sinh(k * (x - x0)) /
                  (2. * (np.cosh(k * (x - x0)) + 1.)**2) )
+
+
+def d2logistic_offset_p(x, p):
+    """
+    Wrapper function for :func:`~pyeee.general_functions.d2logistic_offset`: `d2logistic_offset(x, *p)`.
+    """
+    return d2logistic_offset(x, *p)
 
 
 # -----------------------------------------------------------
@@ -364,28 +375,9 @@ def logistic2_offset(x, L1, k1, x01, L2, k2, x02, a):
 
 def logistic2_offset_p(x, p):
     """
-    Double logistic function with offset:
-
-        `L1/(1+exp(-k1(x-x01))) - L2/(1+exp(-k2(x-x02))) + a`
-
-    where parameters `L1/2`, `k1/2` and `x01/2` are given as one iterable, i.e.
-
-        `p[0]/(1+exp(-p[1](x-p[2]))) - p[3]/(1+exp(-p[4](x-p[5]))) + p[6]`
-
-    Parameters
-    ----------
-    x : array_like
-        Independent variable to evalute logistic function
-    p : iterable
-        Iterable of length 6 with maximum, steepness, inflection point of first logistic function,
-        maximum, steepness, inflection point of second logistic function, and offset of double logistic function
-
-    Returns
-    -------
-    float or ndarray
-        Double Logistic function with offset at `x`
+    Wrapper function for :func:`~pyeee.general_functions.logistic2_offset`: `logistic2_offset(x, *p)`.
     """
-    return logistic2_offset(x, p[0], p[1], p[2], p[3], p[4], p[5], p[6])
+    return logistic2_offset(x, *p)
 
 
 # -----------------------------------------------------------
@@ -426,6 +418,13 @@ def dlogistic2_offset(x, L1, k1, x01, L2, k2, x02, a):
     """
     return ( k1 * L1 / (2. * (np.cosh(k1 * (x - x01)) + 1.)) -
              k2 * L2 / (2. * (np.cosh(k2 * (x - x02)) + 1.)) )
+
+
+def dlogistic2_offset_p(x, p):
+    """
+    Wrapper function for :func:`~pyeee.general_functions.dlogistic2_offset`: `dlogistic2_offset(x, *p)`.
+    """
+    return dlogistic2_offset(x, *p)
 
 
 # -----------------------------------------------------------
@@ -470,6 +469,13 @@ def d2logistic2_offset(x, L1, k1, x01, L2, k2, x02, a):
                  (2. * (np.cosh(k2 * (x - x02)) + 1.)**2) )
 
 
+def d2logistic2_offset_p(x, p):
+    """
+    Wrapper function for :func:`~pyeee.general_functions.d2logistic2_offset`: `d2logistic2_offset(x, *p)`.
+    """
+    return d2logistic2_offset(x, *p)
+
+
 # -----------------------------------------------------------
 
 if __name__ == '__main__':
@@ -504,3 +510,13 @@ if __name__ == '__main__':
     # # 0.5
     # dlogistic2_offset(1.,  1., 2., 1.,  2., 2., 1.,  1.)
     # # -0.5
+    # print(np.around(d2logistic(1., 1., 2., 2.),4))
+    # # 0.3199
+    # print(np.around(d2logistic_offset(1., 1., 2., 2., 1.),4))
+    # # 0.3199
+    # print(np.around(d2logistic2_offset(1., 1., 2., 2.,  2., 2., 2.,  1.),4))
+    # # -0.3199
+    # print(np.around(curvature(1., dlogistic_offset, d2logistic_offset, 1., 2., 2., 1.),4))
+    # # 0.2998
+    # print(np.around(curvature(1., dlogistic_offset_p, d2logistic_offset_p, [1., 2., 2., 1.]),4))
+    # # 0.2998